Crate s502_ln

Modules

formats
In addition to the Tilt and assembly files, the S502 Toolchain has two internal file formats. The assembler outputs object files which are expected to have teh extention .65o, and it has the following format:

input

linker

script
Linker scripts allow you to tell the linker how to organize the sections in the object files.

Functions

main