Module s502_ln::formats

In addition to the Tilt and assembly files, the S502 Toolchain has two internal file formats.

Object files contain code symbol information for the linker to combine into a single binary. Symbol files contain labels and their addresses in memory extracted from previously linked binaries.

The assembler outputs object files which are expected to have the extension .65o, and it has the following format (all ASCII bytes strings are null-terminated):

Object Header

num_sections: u32

followed by sections:

Section Header

name: 32 ASCII bytes
size: u32
num_parents: u32
num_references: u32

name is a simple identifier. size refers to the number of bytes in its payload. num_parents is the number of parent labels in the section, and num_references is the number of references to labels in the section.

Following the section header is a flattened tree of parent and child labels, and they appear in order of their offset into the section. The first label is guaranteed to be a parent with the form:

Parent Label

name: 32 ASCII bytes
num_children: u32
offset: u32
visibility: u32

offset is the offset into the section's payload that this label is located at. visibility dictates from where this label may be referred to:

1 -> from anywhere in this object
2 -> globally from any object

After a parent label is a number of child labels with the form:

Child Label

name: 32 ASCII bytes
offset: u32
visibility: u32

offset is the same as in the parent label, though visibility is slightly changed:

0 -> May only be referred to from under the same parent. That is,
  let ref = offset of reference to this label
  let parent = offset of this label's parent
  let next = offset of the next parent label
  (parent <= ref < next) must hold in order to resolve the reference.
1 -> from anywhere in this object
2 -> globally from any object

After the child labels of a parent is the next parent label.

After all of the labels is the reference block. Each reference has the form:

Reference

referred: 64 ASCII bytes
offset: u32
which_byte: u16
branch: u16

referred is the fully qualified name of the label being referred to. If it refers to a child label then it will follow the parent label and a period, for example reset.skip_config. offset is the offset into the section at which to put the address of the referred label. which_byte dictates which byte from the label's address to put into offset:

0 -> both
1 -> high byte
2 -> low byte

branch tells if the opcode before it was a branch instruction. If its value is 1 then which_byte is ignored and the difference between reference's offset + 1 and the label's address is inserted, and it must fit in one signed byte.

After the references is the section's payload. Only sect_header.size bytes are significant but it is padded to a 4 byte boundary. The number of bytes to read may be calculated size + ((4 - (size & 3)) & 3).

Symbol Table

The linker may output symbol tables while linking objects together. This is convenient for resolving references to binaries that were linked separately and loaded elsewhere in memory without having to link against the actual object. The symbol table has the form:

Symbol Table Header

num:symbols: u32

###Symbol

name: 64 ASCII bytes
address: u32

Structs

Label

Reference

RelocGroup
A group of section relocations.

Section
A section as read from the object file.

Enums

ByteSelect

Visibility

Constants

_IMPL_NUM_FromPrimitive_FOR_ByteSelect

_IMPL_NUM_FromPrimitive_FOR_Visibility