Module s502_ln::script

Linker scripts allow you to tell the linker how to organize the sections in the object files.

// single line comments are allowed
0x1000 text

This indicates that the text section is expected to be loaded at address 0x1000 in memory. If no start address is give at the beginning of the script then the default is 0.

data 0x100

Here the data section immediately follows text. The 0x100 allows the section to be a maximum of 0x100 bytes.

one two 0x200

Here the sections one and two combined are allowed 0x200 bytes. two immediately follows one.

0x0f00 something

If the something section is larger than 0x100 bytes then it will overwrite the beginning of text because it is listed later in the script. You can set a size limit as in the data section if you want to prevent that. Because it is at the earliest address, it will be at offset 0 in the output binary file.

Enums

Token
The tokens recognized in the linker script.

Functions

read_group
Reads a line into a relocation group.

read_script
Reads a linker script into relocation groups.