Module s502_ln::formats

In addition to the Tilt and assembly files, the S502 Toolchain has two internal file formats. The assembler outputs object files which are expected to have teh extention .65o, and it has the following format:

Object Header

num:sections: u32

followed by sections:

Section Header

Structs

Label

Reference

Relocation

Section

Enums

ByteSelect

Visibility